Apply wrappers during RubyLandProxy creation.
This allows pure-ruby code to control how different JS values are
presented to the rest of ruby-land. Such wrappers provide for the most
common case, but there's also a need to be able to define a conversion
to be applied each time the proxy is required, too. And then we need
corresponding functionality in js-land.